% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSubDisc-package.R
\docType{package}
\name{rSubDisc-package}
\alias{rSubDisc}
\alias{rSubDisc-package}
\title{rSubDisc: Wrapper for Java subgroup discovery library SubDisc}
\description{
SubDisc is a tool which, given a dataset, performs Subgroup Discovery. This is a form of Supervised Descriptive Local Pattern Mining. For more info https://github.com/SubDisc/SubDisc
}
\author{
\strong{Maintainer}: Jeremie Gobeil \email{jeremie.gobeil@gmail.com} (\href{https://orcid.org/0000-0002-8920-4127}{ORCID})

}
\keyword{internal}
