% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subdisc.R
\name{subgroupdiscovery}
\alias{subgroupdiscovery}
\title{Search a dataset for subgroup}
\usage{
subgroupdiscovery(
  src,
  targetColumn,
  targetValue = NULL,
  targetType = .subdisc.TargetType("SINGLE_NOMINAL"),
  qualityMeasure = .subdisc.QualityMeasure("CORTANA_QUALITY"),
  qualityMeasureMinimum = NULL,
  searchDepth = NULL,
  minimumCoverage = NULL,
  maximumCoverageFraction = NULL,
  maximumSubgroups = NULL,
  maximumTime = NULL,
  searchStrategy = .subdisc.SearchStrategy("BEAM"),
  nominalSets = NULL,
  numericOperatorSetting = .subdisc.NumericOperatorSetting("NORMAL"),
  numericStrategy = .subdisc.NumericStrategy("NUMERIC_BEST"),
  searchStrategyWidth = NULL,
  nrBins = NULL,
  nrThreads = 1
)
}
\arguments{
\item{src}{The file to load}

\item{targetColumn}{The column number to use as the target}

\item{nrThreads}{}
}
\value{

}
\description{
Search a dataset for subgroup
}
