% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subdisc.R
\name{.subdisc.single_numeric.explained_variance}
\alias{.subdisc.single_numeric.explained_variance}
\title{Title}
\usage{
.subdisc.single_numeric.explained_variance(
  src,
  targetColumn,
  qualityMeasureMinimum = 0.1,
  searchDepth = 1,
  minimumCoverage = 2,
  maximumCoverageFraction = 1,
  maximumSubgroups = 1000,
  maximumTime = 1000,
  searchStrategy = .subdisc.SearchStrategy("BEAM"),
  nominalSets = FALSE,
  numericOperatorSetting = .subdisc.NumericOperatorSetting("NORMAL"),
  numericStrategy = .subdisc.NumericStrategy("NUMERIC_BEST"),
  searchStrategyWidth = 10,
  nrBins = 8,
  nrThreads = 1
)
}
\arguments{
\item{nrThreads}{}
}
\value{

}
\description{
Title
}
