% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subdisc.R
\name{subgroupdiscovery}
\alias{subgroupdiscovery}
\title{Search a dataset for subgroup}
\usage{
subgroupdiscovery(
  src,
  targetColumn,
  targetValue = NULL,
  targetType = "SINGLE_NOMINAL",
  qualityMeasure = "CORTANA_QUALITY",
  qualityMeasureMinimum = 0.1,
  searchDepth = 1,
  minimumCoverage = 2,
  maximumCoverageFraction = 1.0,
  maximumSubgroups = 1000,
  maximumTime = 1000,
  searchStrategy = "BEAM",
  nominalSets = FALSE,
  numericOperatorSetting = "NORMAL",
  numericStrategy = "NUMERIC_BEST",
  searchStrategyWidth = 10,
  nrBins = 8,
  nrThreads = 1
)
}
\arguments{
\item{src}{The file to load}

\item{targetColumn}{The column number to use as the target}

\item{nrThreads}{The number of CPU threads SubDisc is allowed to create.}
}
\value{

}
\description{
Search a dataset for subgroup
}
